domain/item/item_test.go
package item_test

import (
	"testing"

	"github.com/dannegm/spendzer-service/domain/item"
)

func TestItem_NewItem(t *testing.T) {
	// Build our needed testcase data struct
	type testCase struct {
		test        string
		name        string
		expectedErr error
	}
	// Create new test cases
	testCases := []testCase{
		{
			test:        "Empty Name validation",
			name:        "",
			expectedErr: item.ErrInvalidItem,
		}, {
			test:        "Valid Name",
			name:        "Banana",
			expectedErr: nil,
		},
	}

	for _, tc := range testCases {
		// Run Tests
		t.Run(tc.test, func(t *testing.T) {
			// Create a new customer
			_, err := item.NewItem(tc.name)

			// Check if the error matches the expected error
			if err != tc.expectedErr {
				t.Errorf("Expected error %v, got %v", tc.expectedErr, err)
			}
		})
	}
}
